feat: Refactor Request and Response classes
