+++ src/main/java/co/novu/api/common/NotificationGroup.java
@@ -1,4 +1,4 @@
-package co.novu.api.workflows.pojos;
+package co.novu.api.common;
 
 import lombok.Data;
 
@@ -9,4 +9,7 @@ public class NotificationGroup {
     private String _organizationId;
     private String _environmentId;
     private String _parentId;
+    private String createdAt;
+    private String updatedAt;
+    private String __v;
 }
